requirements.txt
asgiref==3.4.1
astroid==2.4.2
attrs==21.2.0
beautifulsoup4==4.10.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.2.5
django-crispy-forms==1.11.2
django-filter==2.4.0
django-url-filter==0.3.15
django-widgets-improved==1.5.0
djangorestframework==3.12.2
enum-compat==0.0.3
gunicorn==20.1.0
idna==2.10
imgkit==1.2.2
isort==5.7.0
lazy-object-proxy==1.4.3
lxml==4.7.1
Markdown==3.3.3
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.22.2
opencv-contrib-python==4.5.5.62
packaging==21.3
pdf2image==1.16.0
pdfkit==0.6.1
Pillow==9.0.1
psycopg2-binary==2.9.1
pylint==2.6.0
pyparsing==3.0.7
pyrsistent==0.17.3
pytesseract==0.3.9
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.3.1
sqlparse==0.4.1
timeago==1.0.15
toml==0.10.2
urllib3==1.26.4
wincertstore==0.2
wrapt==1.12.1
xmltodict==0.12.0
zpl==0.1.8
